refactor: 스토리 작성 페이지에서 핸들링하는 데이터를 도메인별로 분리 #6
