--- .env
@@ -1,2 +1,0 @@
-# Default Environment Variables
-# 공통 설정이 필요한 경우 여기에 추가
